revert not needed changes
