+++ packages/gene/src/ioc/react.ioc.tsx
@@ -108,3 +108,19 @@ export function useCallbackInjection<T, R = void>(
     return fallback;
   }
 }
+
+export function withIoc<Props extends Record<string, any>>(
+  getContainer: (props?: Props) => interfaces.Container
+) {
+  return (Page: React.ComponentType<Props>) => {
+    return (props: Props) => {
+      const container = getContainer(props);
+
+      return (
+        <Provider container={container}>
+          <Page {...props} />
+        </Provider>
+      );
+    };
+  };
+}
